Utility to support SAFT and a unit test for it.
saft/flash_handler.py is a helper utility used by the SAFT\
(semi automated firmware test).

It provides the ability to load a firmware file (or read it
from the hardware when running on the target), verify the
file's integrity and modify certain section of the file to
introduce verification errors.

safe/test_flashrom_handler.py is a unit test for this
utility. It can run on host (in native and chroot
environments) and on the target.

x86-generic/flashrom_util.py is modified to suppress the
console output generated by the 'flashrom' utility when
operating on the target.

BUG=chromium-os:1994

TEST=See below

This is a sample unit test invocation case (with firmware
image placed in file zgax64.fd in the local directory):

vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
(saft $) PATH=${PATH}:../../vboot_reference/build/utility \
 ./test_flashrom_handler.py \
 ../../vboot_reference/tests/devkeys/root_key.vbpubk \
 zgax64.fd
----------------------------------------------------------------------
Ran 2 tests in 0.643s

OK
(saft $)
^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

The

Review URL: http://codereview.chromium.org/2834065