First draft SAFT implementation.
This CL introduces the semiautomated firmware test (SAFT)
framework, as described in
https://docs.google.com/a/google.com/document/edit?id=1E1b_73IQjSiRsznSh_YDhv5i3mn__yDeZIGhXwY8rhA&hl=en#

The required parameters of ./saft_utility are
 --image_file - name of the file containing the firmware
   image which needs to be tested
 --pub_key - name of the file contaning public key block
   used to verify both currently running and new firmware
   images

saft_utility must run off the flash stick. On startup
in verifies integrity of both current and new FW images,
prepares the environment, and creates the upstart script
(/etc/init/fw_test.conf) on three partitions: both root
partitions on the hard drive/SSD and the recovery partition
on the flash stick.

Then the test restarts the machine.

The upstart script makes sure that saft_utility is invoked
after each reboot with the --next_step command line
parameter.

The expected test states are enumerated in
saft_utility:test_state_sequence table. On each step the
test verifies the proper state on the machine, performs
the necessary action and restarts the machine.

The test ends when there is no action in the current record
of the state machine table or in case an error happens.

The test progress is accumulated in the log file stored on
the stateful partition on the flash stick (in
/var/.fw_test/fw_test_log.txt)

BUG=chromium-os:1994

Test=see below

1. Unit test ran on the target:

localhost fw_test # ./test_saft_utility.py
...
----------------------------------------------------------------------
Ran 3 tests in 0.894s

OK
localhost fw_test # cd saft

localhost saft # ./test_flashrom_handler.py /etc/keys/root_key.vbpubk
..
----------------------------------------------------------------------
Ran 2 tests in 11.950s

OK
localhost saft #

2. The actual SAFT ran on the target:

Place the fwirmware/ directory contents on a flash stick, place the new flashrom image into a file in the local directory (./ZGA_010b.fd) and run the test:

./saft_utility  --pub_ke=/etc/keys/root_key.vbpubk --ima=./ZGA_010b.fd

The test causes the machine to restart several times, in the end the following is seen in the log file (/var/.fw_test/fw_test_log.txt on the first partition of the flash stick):

12:51:40 PM: Automated firmware test log generated on Jul 25 2010
12:51:40 PM: Original boot state 1:1:0
12:51:46 PM: Executing df /tmp/sdb3/usr/bin/fw_test
12:51:46 PM: Executing blkid
12:51:46 PM: Executing rootdev
12:51:46 PM: Executing mount
12:51:46 PM: Handling //etc/init/fw_test.conf
12:51:46 PM: Executing rootdev
12:51:46 PM: Executing mount
12:51:46 PM: Handling /tmp/sdb3/etc/init/fw_test.conf
12:51:46 PM: Executing rootdev
12:51:46 PM: Executing mount
12:51:46 PM: Executing mount /dev/sda5 /tmp/tmpuC0YjA
12:51:46 PM: Handling /tmp/tmpuC0YjA/etc/init/fw_test.conf
12:51:46 PM: Executing umount /tmp/tmpuC0YjA
12:51:46 PM: program the new image
12:52:18 PM: restart
12:52:18 PM: Executing reboot
12:52:44 PM: Rebooted into state 1:1:0 on step 0
12:52:44 PM: calling <function SetTryFwB at 0x6f744fb4>
12:52:44 PM: Requesting restart with FW B
12:52:44 PM: Executing reboot_mode --try_firmware_b=1
12:52:44 PM: Executing reboot
12:53:07 PM: Rebooted into state 1:2:0 on step 1
12:53:08 PM: calling <function <lambda> at 0x6f76d614>
12:53:08 PM: Executing reboot
12:53:28 PM: Rebooted into state 1:1:0 on step 2
12:53:29 PM: calling <function CorruptFirmware at 0x6f74717c> with parameter a
12:54:05 PM: Executing reboot
12:54:25 PM: Rebooted into state 1:2:0 on step 3
12:54:25 PM: calling <function RestoreFirmware at 0x6f6c70d4>
12:54:25 PM: Restoring to original firmware
12:54:25 PM: Executing flashrom -w /tmp/tmp_dbSPi/var/.fw_test/flashrom.bak
12:54:57 PM: Executing reboot
12:55:16 PM: Rebooted into state 1:1:0 on step 4
12:55:17 PM: Removing upstart scripts
12:55:17 PM: Executing df /tmp/sdb3/usr/bin/fw_test
12:55:17 PM: Executing blkid
12:55:17 PM: Executing rootdev
12:55:17 PM: Executing mount
12:55:17 PM: Executing mount -o remount,rw /dev/root
12:55:17 PM: Handling //etc/init/fw_test.conf
12:55:17 PM: Executing rootdev
12:55:17 PM: Executing mount
12:55:17 PM: Handling /tmp/sdb3/etc/init/fw_test.conf
12:55:17 PM: Executing rootdev
12:55:17 PM: Executing mount
12:55:17 PM: Executing mount /dev/sda5 /tmp/tmpgfrUqx
12:55:17 PM: Handling /tmp/tmpgfrUqx/etc/init/fw_test.conf
12:55:17 PM: Executing umount /tmp/tmpgfrUqx
12:55:18 PM: we are done!

Review URL: http://codereview.chromium.org/3061012