Include root fs information in the boot state vector.
This CL adds two elements to the boot vector describing the
ChromOS system boot state.

The second last element is number 1 or 0 (depending if the
root file system is mounted on a removable or non-removable
device, respectively). The last number is the digit of the
root fs device. The root FS device hardcoded to match the
device the kernel was loaded from, so the combinations of
the last two elements are as follows:

0:3 => non removable, kernel A
0:5 => non removable, kernel B
1:3 => removable, recovery kernel

BUG=BUG=chromium-os:4801
TEST=see http://codereview.chromium.org/2868098/show
for test decription

Review URL: http://codereview.chromium.org/3036039