Introduce gpt tests in SAFT.
This CL introduces infrastructure which allows SAFT to
access gpt (GUID partition table) information using the
cgpt utility.

There are may aspects of gpt management which can and should
be tested, this CL just creates a module which provides
a means of retrieving gpt information and changing it for
the test purposes, and adds one gpt test as an example.

A new module (cgpt_handler.py) is introduced along with the
unit test (test_cgpt_handler.py). See module docstrings
for details of operations.

saft_utility.py is being cleaned up as follows:

- imported modules arranged alphabetically, global then
  local

- if an attempt to execute a step results in a failure, call
finish_saft(False)to remove the saft 'follow up' script.
This is needed because gpt tests are more involved and can
not be reduced to checking of the boot vector value.

The target comes up with the proper boot vector, but the
cgpt reported values could still be wrong. In case cgpt
output does not match expected values, the test terminates.

- in case of error, do not call chros_if.shut_down(). This
will prevent moving the log file into /var directory and
removal of the test temp directory (/var/.fw_test on the
flash device). This will allow to examine the state of the
failed test in more details. The last step error messages
are still available in /tmp/saft.log

Also, a gpt related test case is being added:

- set the kernel tries attribute to 15 and observe it to
be changed to 14 after the reboot.

More gpt test can be added following this example.

Change-Id: Idc5a36baed8a8a0241dbeca8f62abe75e2136ba1

BUG=chromium-os:1994
TEST=see below

A new image was built and the test was run several times.
All test pass, the following section is included in the log file:

06:17:03 PM: calling <bound method FirmwareTest.cgpt_test_start of <__main__.FirmwareTest object at 0x6f69c62c>>
06:17:03 PM: Executing cgpt show /dev/sda
06:17:03 PM: Executing cgpt add -i 2 -S 0 -P 10 -T 15 /dev/sda
06:17:03 PM: Executing cgpt add -i 4 -P 9 -T 15 /dev/sda
06:17:03 PM: Executing reboot
06:17:16 PM: Executing cgpt show /dev/sda
06:17:16 PM: Executing rootdev -s
06:17:16 PM: Rebooted into state 1:1:0:0:3 on step 10
06:17:16 PM: startup log contents:
06:17:16 PM: Found sdc after 2 seconds
06:17:16 PM: calling <bound method FirmwareTest.cgpt_test_1 of <__main__.FirmwareTest object at 0x6f71762c>>
06:17:16 PM: Executing cgpt show /dev/sda
06:17:16 PM: Executing reboot
06:17:30 PM: Executing cgpt show /dev/sda
06:17:30 PM: Executing rootdev -s
06:17:30 PM: Rebooted into state 1:1:0:0:3 on step 11

Review URL: http://codereview.chromium.org/3385007