Workaround for lost CMOS flags in boot.
The firmware updater sets the try-firmware-B flag in the CMOS, but sometimes
the CMOS loses its state and the flag gets cleared. This adds a flag file to
the stateful partition so that if the next boot discovers that we've booted
firmware A and we wanted to boot B, it sets the CMOS flag, clears the flag
file, and reboots quickly.

Change-Id: Ibc372004cc3d991ce0d92b7da2aa689d3ba04692

BUG=chrome-os-partner:1563
TEST=manual

NOTE: Be sure that the autoupdater is not running. The easiest way to
prevent it is to disable all network connections, both wired and wireless,
before starting this test.

On Mario, log in as root. Verify that you've booted from firmware A by
running

  cat /sys/devices/platform/chromeos_acpi/BINF.1

It should contain "1". If it contains "2", reboot, and try again.

Now, run this:

  touch /mnt/stateful_partition/.need_firmware_update
  touch /mnt/stateful_partition/.need_firmware_tryb
  reboot_mode --try_firmware_b=0
  reboot

You should see the machine start to boot, stop, then reboot normally. Log in
as root again, and run

  cat /sys/devices/platform/chromeos_acpi/BINF.1

It should contain "2", indicating that you've booted with firmware B. You
should also run

  ls -a /mnt/stateful_partition

and verify that both .need_firmware_update and .need_firmware_tryb are no
longer present.

Review URL: http://codereview.chromium.org/4339002