firmware updater: use new flashrom command syntax (-p internal:bus)
The new option "-p internal:bus" of flashrom is cleaner and prevents race
condition for scripting flashrom, and also helps in dropping dependency of
iotools.

The ACPI initialization is not required for a long time so it's also cleaned up in this CL. Definition is still kept to work with legacy code in some per-board customization.

BUG=chromium-os:10685
TEST=manually invoked updater on Mario and it seems running fine.

Change-Id: I5b517278054022fc92cfa61a9a260a7832954a98

Review URL: http://codereview.chromium.org/6377008