chromeos-firmware: replace ACPI/GPIO by crossystem command
Crossystem provides a better way to access all the low level system resources,
so both ACPI/GPIO stuff can be wrapped by crossystem.

Although this is in the legacy firmware updater,
since it will still be used for CR-48 for a while,
we still need to fix it.

BUG=chrome-os-partner:2420, chromium-os:12327
TEST=dry-run on CR-48.

Change-Id: I12ccfe8a7b1febb2d8f96b0efac2903192042976

Review URL: http://codereview.chromium.org/6646014