firmware-updater: Preserve VPD in more fault-tolerance way.
When the system active firmware has FMAP, we should trust it and preserve VPD
from that; if the firmware (ex, non-chrome) does not have FMAP, use the FMAP
from target image.

And for all cases, check if the VPD partition is valid; otherwise reset it.

BUG=chrome-os-partner:7185
TEST=Prepare two incompatible firmware: (ex, H2O v.s. H2C or *boot)
     that VPD is located in different location (examples in 7185);
     Set some VPD values (vpd -i RO_VPD -s a=b);
     Run chromeos-firmwareupdate --mode=factory_install;
     Then, reboot and verify if VPD is preserved correctly (vpd -i RO_VPD -l)

Change-Id: I2f6cfc4275c25c342d6fc93fe5c2c14fc4a89fd7
Reviewed-on: https://gerrit.chromium.org/gerrit/12974
Reviewed-by: Randall Spangler <rspangler@chromium.org>
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>