Fix firmware packing for EC + BIOS version string containing slash char.
If '/' is present in the EC or BIOS version string, pack_firmware.sh
will fail because '/' is the sed delimiter character. Here we change the
delimiter character to ascii 001 (SOH), which is not very likely to
appear in our strings. The delimiter changing feature of sed is
documented in the sed manual:

http://www.gnu.org/software/sed/manual/sed.html#Addresses

BUG=chromium-os:33969
TEST=emerge-boardname chromeos-firmware-boardname

Change-Id: I789b6bb735ba22d1ac25e5178d9cd55902563ab6
Reviewed-on: https://gerrit.chromium.org/gerrit/31668
Tested-by: Shawn Nematbakhsh <shawnn@chromium.org>
Reviewed-by: Jon Salz <jsalz@chromium.org>
Commit-Ready: Shawn Nematbakhsh <shawnn@chromium.org>