chromeos-firmwareupdate: Add support for updating PD FW
This adds support for updating the PD firmware image into the
existing updater4 script.  If a platform firmware package bundles
a PD firmware it will be updated with the same rules as the EC.

It is a transparent noop for platforms that do not bundle a PD
firmware image.

BUG=chrome-os-partner:31130
TEST=generate bundle with emerge-samus chromeos-firmware-samus
ensure BIOS/EC/PD update as expected with different modes

Change-Id: I09ec9748b7d203a586ae8f9ba7cadf04a3f4c47f
Signed-off-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/213288
Reviewed-by: Hung-Te Lin <hungte@chromium.org>