updater: Fix legacy update rules.
The LEGACY_UPGRADE_WHITELIST is a list of hashes that "must be
upgraded". That means when we are checking if a legacy update can be
applied, we should only check if "current" (not target) matches the
white list.

Previously the code has done such check by 'if [ "$legacy_hash" = "$hash" ] &&
 cros_fw_is_equal_slot "$TYPE_MAIN" "$SLOT_LEGACY"' which has two
problems:

1. cros_fw_is_equal_slot's correct name now is crosfw_is_equal_slot.
2. The function compares "current" and "target", which does not help
   deciding if we should update.

Take this for example: WHITE_LIST=A, CURRENT=B, UPDATE_TARGET=C:
 A != B => Can't update (user has installed something else)
 A = B != C => Need to update.
 A = B = C => Should not happen because that means target itself is also
              broken (flashrom will also ignore this).

So we know it should be fine to simply compare current (slot_hash) and
white list blob.

BUG=none
TEST=none

Change-Id: I06bf464d472dca889b2c0166320bbd17d23f57aa
Reviewed-on: https://chromium-review.googlesource.com/329468
Commit-Ready: Hung-Te Lin <hungte@chromium.org>
Tested-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-by: Colin Rice <clr@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>