pack_stub: Don't use a timestamp on the version files
Use a standard timestamp for the version files so that we get the same
exact sharball each time (with the same input files). Otherwise the
changing timestamps creates small differences between each successive run
of the ebuild, making it harder to verify that nothing has changed.

This is safe because this script uses the shar -m flag when unpacking,
meaning that timestamps are ignored ayway.

BUG=chromium:690573
TEST=emerge-reef chromeos-firmwareupdate
with debugging to show the version file timestamps:
-rw-r--r-- 1 sjg portage      549 Jan  1 00:00 VERSION
-rw-r--r-- 1 sjg portage      605 Jan  1 00:00 VERSION.md5

Change-Id: I1f702c2f43c6a0615a16aa48a532af0ceadf7a5b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/440645
Reviewed-by: Hung-Te Lin <hungte@chromium.org>