Make a start on rewriting pack_firmware.sh in Python
This is a long script (~370 lines) and would be easier to maintain in
Python than bash. Also we plan some very significant enhancements for
unified builds, which would be painful to implement in bash.

Start by parsing the arguments. There is one difference, which is that you
must use --no-merge_bios_rw_image to disable that option. This is not
used anywhere in Chrome OS, and help will provide info for anyone that
tries it.

BUG=chromium:690573
CQ-DEPEND=442673
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py

Change-Id: I3da91734279ac96566a8698e03108766b50e16ae
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/440649