Adjust the presubmit to use PRESUBMIT_FILES
This script (from login-manager) does not do what I want it to do. It
looks like I can use an environment variable to find out what files are
in the commit, so do that instead.

BUG=chromium:690573
TEST=try uploading changes with and without modifying pack_firmware.sh in
the commit. See that it detects any change correctly and prints an error
when detected.

Change-Id: I0590c5c9c160283476202499f4abf534a40d3c23
Reviewed-on: https://chromium-review.googlesource.com/450890
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>