Add a functional test for firmware update
The existing unit test is useful for finding simple problems. But there is
value in having an larger test which actually runs the firmware update to
make sure it behaves as expected.

To make this work we need a firmware image. We also need to mock out some
of the tools:

- crossystem - used to display info about the system we are running on
- flashrom - used to access the SPI firmware
- vpd - used to access the VPD area of firmware

BUG=chromium:690573
TEST=FEATURES=test emerge-reef --nodeps  chromeos-firmware-reef
Outside chroot:
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I16ad520573302a8f9a713dab2ba8400690ff03c7
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/450855
Reviewed-by: Dan Erat <derat@chromium.org>