Adjust make_exec_temp to remount only when running as root
When running the firmware update from a functional test on a host machine
we cannot remount /tmp. Nor do we need to as it has execution permission
already.

Adjust the script to do the remount only when running as root.

BUG=chromium:690573
TEST=Run firmware update script as root on reef; check that it completes
correctly:
$ sh chromeos-firmwareupdate --debug
bind and remount for allowing execution ...
Start running script: ./updater4.sh --debug
...

Change-Id: Id91be4d26493ce69d46b4fba41681cd4b0945349
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/450854