pack_firmware.sh: Don't embed the temporary directory name in VERSION
When RW firmware is created from RO firmware, the shellball does not
produce the same output each time, since the name of the temporary
directory ends up in the VERSION file. Fix this by replacing it with a
constant string.

BUG=chromium:690573
TEST=emerge-whirlwind --nodeps chromeos-firmware-whirlwind
with https://chromium-review.googlesource.com/c/442790/ applied
With other changes, this passes.

Change-Id: Iba70fb40b7c06e965be37dbf5b4cc6b406ec1673
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/452777
Reviewed-by: Dan Erat <derat@chromium.org>