pack_stub: Don't mess with the whitespace
The pack_stub updates the version information but the regexes are a bit
loose with their treatment of whitespace. Update them to preserve the
whitespace as is, to avoid adding one more variable in the script's
behaviour.

BUG=chromium:690573
TEST=emerge-whirlwind --nodeps chromeos-firmware-whirlwind
with https://chromium-review.googlesource.com/c/442790/ applied
With other changes, this passes.

Change-Id: Icda634d168903334cb54fde44878530e07f016ab
Reviewed-on: https://chromium-review.googlesource.com/452786
Commit-Ready: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>