pack_firmware.py: Use a known timestamp for created RW firmware
When creating RW firmware from RO firmware, use the same timestamp. This
allows us to get the same output each time the script is run.

Add a test for this as well. To make this work, provide a test-only option
to keep the temporary directories around.

BUG=chromium:690573
TEST=emerge-whirlwind --nodeps chromeos-firmware-whirlwind
with https://chromium-review.googlesource.com/c/442790/ applied
With other changes, this passes.

Change-Id: Ib985b62a4715c33296404f473d9e384489287067
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/452781
Reviewed-by: Dan Erat <derat@chromium.org>