pack_firmware.py: Correct logic in _CheckRwFirmware()
The logic for checking for an RW image is actually inverted. Fix it. Also
update the function comment which is silent on this.

Use a mock to handle the testing since we don't want to have two separate
versions of the real image just for testing.

BUG=chromium:690573
TEST=emerge-whirlwind --nodeps chromeos-firmware-whirlwind
with https://chromium-review.googlesource.com/c/442790/ applied
With other changes, this passes.

Change-Id: I84d91442c1aad1fef55743a34cf7f2e324813682
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/452782
Reviewed-by: Dan Erat <derat@chromium.org>