pack_firmware.py: Don't embed the temporary directory name in VERSION
When RW firmware is created from RO firmware, the shellball does not
produce the same output each time, since the name of the temporary
directory ends up in the VERSION file. Fix this by replacing it with a
constant string.

BUG=chromium:690573
TEST=emerge-whirlwind --nodeps chromeos-firmware-whirlwind
with https://chromium-review.googlesource.com/c/442790/ applied
With other changes, this passes.

Change-Id: Ib7c2fac53f6420d705f6cb06bc1f17c6006e2902
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/452785
Reviewed-by: Hung-Te Lin <hungte@chromium.org>