pack_stub/pack_dist: Update on-device scripts for unified builds
With unified builds the firmware update has an extra step at the start
since it needs to support updating multiple models. We need to check the
device's model at the start.

This task is performed by the new cros_set_unibuild_vars() function. When
called it reads the model name from the hardware ID (HWID). It uses this
model name to find the file containing shell variable assignments for the
model. The file is created (in a later CL) from setvars_template, with
appropriate values substituted in the 'REPLACE_...' strings.

Once this file is executed, we are in the same position as for a
single-model firmware update: we know the images and versions to be used
by the update.

From there the firmware update proceeds as normal.

Therefore the only change needed in the updater*.sh scripts is to call
this new cros_set_unibuild_vars() function at the start. Update
updater4.sh to do this so we can support firmware update with unified
builds on reef.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I7c6f67851292ca582ca8e337e4d0b18c1313feec
Signed-off-by: Simon Glass <sjg@chromium.org>