pack_stub: Add a UNIBUILD variable
The firmware update script with unified builds need to cope with selecting
the correct images based on the model name of the device we are running
on. It is no-longer enough to have a single image of each type (BIOS, EC)
so the variables at the top of pack_stub are not useful. Instead, those
variables are set in separate scripts, one for each model.

As a first step, define a UNIBUILD variable at the top of the script, and
avoid setting the existing variables if that is enabled.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: Iae9b0426bc59db3c9c149cb88d174fbc8dc5ae0e
Signed-off-by: Simon Glass <sjg@chromium.org>