pack_firmware.py: Use a dict to store version info
_CopyFirmwareFiles() both copies the files and decides what to copy. It
handles this by setting the program argument self._args.merge_bios_rw_image
to False when it decides that it cannot in fact merge.

This mirrors the old shell script but is not very nice. It would be better
to make all the decisions up front and then perform the tasks without any
thought.

As a first step towards this, rather than writing directly into the
version information file, collect the information in a dict and write it
at the end. This will allow us to write the information out-of-order in a
future CL.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I835bf8652010e3cc9a8f609adb35aa2b893794fc
Signed-off-by: Simon Glass <sjg@chromium.org>