pack_firmware.py: Move inner control code into a function
With unified builds we need to add files for multiple models into the
same firmware bundle. In preparation for this, move the inner part of the
code into a new _ProcessModel() function and call it with the supplied
program arguments. With unified builds we will call it multiple times
with different values, collected from the master configuration.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I83f1be291d62419723d726a18f160e51261edbdd
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/453379
Reviewed-by: Dan Erat <derat@chromium.org>