pack_firmware.py: Run tests with dash
Chrome OS uses the dash shell and the scripts will use that when they
include the '/bin/sh' shebang at the top. However the chroot uses bash
so the same shebang has a different meaning.

In order to make the test more faithful to the target environment,
add a way to change the shebank to explicitly select dash. This is only
used in tests and provides more confidence that we are not using
bash features in the firmware update scripts.

When this flag is enabled, the script replaces all shebangs with
/bin/dash.

While we are changing _CopyFile(), fix its documentation.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py
Also add a bash-specific construct in pacK_stub and see that the
functional tests now fail.

Change-Id: I341a65df87b39c71498032b569fb207d861773e9
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/453295