pack_firmware.py: Split up _WriteUpdateScript()
With unified builds we need to write multiple 'setvars' scripts each of
which sets the firmware variables for a particular model. The current
code just updates the top of the (single) update script.

In preparation for this, move the code which substitutes variables into
its own function, called by WriteUpdateScript(). In a future CL we can
call it from the unified build code.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I3e2468abfa3d90a1f20eb2fa90b3c53068e4aebc
Signed-off-by: Simon Glass <sjg@chromium.org>