pack_firmware.py: Support unpacking tar files automatically
At present the cros-firmware eclass is responsible for unpacking tar
files. This happens in that script's _unpack_archive() function.

This is somewhat clumsy but with unified builds it becomes a pain since it
means that both the eclass and pack_firmware.py need to read all the
firmware configuration. It is much easier if pack_firmware.py handles this
itself.

Add a way to detect tar files downloaded from bcs:// by the eclass, and
unpack the file within. The directory to unpack into is provided as a
command-line flag, since the eclass will need to control this.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I646671a0b464436a8e72c1e339f8f685d5b57564
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455223
Reviewed-by: Dan Erat <derat@chromium.org>