pack_firmware.py: Support unified builds in the update script
With unified builds we need to set the UNIBUILD variable to 'yes' in the
update structure. Also most of the other variables are not used so we
shoud signal that to avoid confusion.

Adjust the behaviour based on a boolean 'unibuild' parameter. For now this
code is unused since we pass False for unibuild.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: I85432d0716316e44ece3035473d35114d924da3b
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455227
Reviewed-by: Dan Erat <derat@chromium.org>