pack_firmware.py: Generate unified build firmware updates
Now that all the pieces are in place we can add support for generating a
firmware update that works on multiple models.

The shellball is created as normal, but we add a separate directory in it
for each model, and place the required files in there. A 'setvars.sh' file
provides the shell variables for the model. The list of models supported
is supplied with the -m flag on the command line. Information about which
iamges and versions to pack into the update comes from the Chrome OS master
configuration file, provided by the -c flag.

Add a few unit tests and a functional test. The latter runs an update on
two models to check that it is happy with either.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py

Change-Id: Ie3613888ade845e31c4b88f0413eed7d680185a3
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/455231