pack_firmware.py: Preserve time stamps when merging firmware
When merging RW firmware into RO, copy the timestamp over as well. This
matches the behaviour when not merging, and allows us to produce the same
output each time.

We could drop the timestamp copy when creating the RW firmware, since the
eclass does not seem to ever create it without also merging it into RO.
But keep it for consistency.

Adjust the relevant test case to check this. Note that it is changed to
call the firmware packer directly rather than through its main() method.
This allows us to preserve the temporary files for checking in the test.
We still have one test which uses the main() method (testMockedRun).

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py
emerge-chell --nodeps chromeos-firmware-chell
with https://chromium-review.googlesource.com/c/442790/ applied
Also check that taking out each _CopyTimestamp() call causes a different
test failure.

Change-Id: I9d893f2dce0c98c43e430776fbc0d4959f93d713
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/456036
Reviewed-by: Dan Erat <derat@chromium.org>