pack_firmware.py: Restore the 'typo' RW_FWID
It turns out that the read-only firmware ID is stored in a section called
RO_FRID while the read-write firmware ID is stored in a section called
RW_RWID. My attempt to fix this 'typo' was in fact incorrect. Re-fix it.

This inconsistency is quite confusing, but I don't this it is feasible to
change it at this stage.

BUG=chromium:690573
TEST=PYTHONPATH=~/cosarm python pack_firmware_unittest.py
PYTHONPATH=~/cosarm python pack_firmware_functest.py
emerge-chell --nodeps chromeos-firmware-chell
with https://chromium-review.googlesource.com/c/442790/ applied

Change-Id: I2bc57a9c7d336373be1febf56db9f44a46026c3d
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/456037
Reviewed-by: Dan Erat <derat@chromium.org>