pack_stub: Sort files in the shellball
It seems that sometimes the ordering of the files changes between the
shell script and the Python script. Sort the contents to avoid this, since
we should not rely on any particular ordering.

BUG=chromium:703869
TEST=USE=chrome_internal  emerge-caroline-userdebug --nodeps
	chromeos-firmware-caroline
$ less /build/caroline-userdebug/usr/sbin/chromeos-firmwareupdate
Go down to 'This shar contains' and see that the files are sorted.
(This CL makes no difference on my machine but I hope it will changes
things on xiaohuic's)

Change-Id: Ibaa5ad5889fdaeaf74399bcd0a60e9630f9f3f26
Signed-off-by: Simon Glass <sjg@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/457842
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Reviewed-by: Dan Erat <derat@chromium.org>
Tested-by: Xiaohui Chen <xiaohuic@chromium.org>